testのディレクトリを統一させた
